requirements.txt
streamlit
google-generativeai
langchain-community
python-dotenv
googletrans==4.0.0rc1
spacy
sentence-transformers
pypdf
faiss-cpu
beautifulsoup4
